backend/requirements.txt
# Backend
asgiref==3.8.1
asyncio==3.4.3
celery[redis]==5.4.0
Django==5.0.7
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
psycopg==3.2.1
